rename linuxes to elinux
