+++ cmake/build.cmake
@@ -118,9 +118,9 @@ add_executable(${TARGET}
   src/flutter/shell/platform/linux_embedded/plugins/text_input_plugin.cc
   src/flutter/shell/platform/linux_embedded/surface/context_egl.cc
   src/flutter/shell/platform/linux_embedded/surface/egl_utils.cc
-  src/flutter/shell/platform/linux_embedded/surface/linuxes_egl_surface.cc
-  src/flutter/shell/platform/linux_embedded/surface/linuxes_surface.cc
-  src/flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl.cc
+  src/flutter/shell/platform/linux_embedded/surface/elinux_egl_surface.cc
+  src/flutter/shell/platform/linux_embedded/surface/elinux_surface.cc
+  src/flutter/shell/platform/linux_embedded/surface/elinux_surface_gl.cc
   ${DISPLAY_BACKEND_SRC}
   ${WAYLAND_PROTOCOL_SRC}
   ## The following file were copied from:
