+++ src/flutter/shell/platform/linux_embedded/surface/elinux_surface.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/linux_embedded/surface/linuxes_surface.h"
+#include "flutter/shell/platform/linux_embedded/surface/elinux_surface.h"
 
 #include "flutter/shell/platform/linux_embedded/logger.h"
