+++ src/flutter/shell/platform/linux_embedded/window/linuxes_window_x11.h
@@ -7,7 +7,7 @@
 
 #include <memory>
 
-#include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl.h"
+#include "flutter/shell/platform/linux_embedded/surface/elinux_surface_gl.h"
 #include "flutter/shell/platform/linux_embedded/window/linuxes_window.h"
 #include "flutter/shell/platform/linux_embedded/window/native_window_x11.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler.h"
