+++ src/flutter/shell/platform/linux_embedded/window_binding_handler.h
@@ -9,7 +9,7 @@
 #include <variant>
 
 #include "flutter/shell/platform/linux_embedded/public/flutter_linuxes.h"
-#include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl.h"
+#include "flutter/shell/platform/linux_embedded/surface/elinux_surface_gl.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler_delegate.h"
 
 namespace flutter {
